requirements.txt
lupa==2.2
polars==1.12.0
psycopg2==2.9.10
python-dotenv==1.0.1
